+++ samples/sample-express-mongodb/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/express-sample-mongodb
 
+## 2.2.0
+
+### Minor Changes
+
+-   Update dependencies
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/starter-openapi@2.2.2
+    -   @nodeboot/express-server@1.7.0
+    -   @nodeboot/starter-actuator@1.5.0
+    -   @nodeboot/starter-http@3.4.0
+
 ## 2.1.27
 
 ### Patch Changes