+++ starters/actuator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/starter-actuator
 
+## 1.5.0
+
+### Minor Changes
+
+-   Update dependencies
+
 ## 1.4.2
 
 ### Patch Changes