+++ starters/http/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/starter-http
 
+## 3.4.0
+
+### Minor Changes
+
+-   Update dependencies
+
 ## 3.3.1
 
 ### Patch Changes