+++ tests/jest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/jest
 
+## 1.2.13
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/test@1.3.0
+
 ## 1.2.12
 
 ### Patch Changes